requirements.txt
bcrypt==5.0.0
Django==5.0
djangorestframework==3.16.1
dotenv==0.9.9
PyJWT==2.11.0
